fix Time data format

+++ app/src/main/java/com/example/fafcalculator/app/model/Result.kt
@@ -3,6 +3,6 @@ package com.example.fafcalculator.app.model
 data class Result(
     val sacu: Int,
     val massIncome: Int,
-    val time: Float,
+    val time: Int,
     var best: Boolean
 )
